email_automation_project/requirements.txt
python-dotenv
email
smtplib
imaplib